backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
openpyxl
chromadb
google-generativeai
sentence-transformers
python-multipart